# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PromptVault API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{slug}:
    get:
      operationId: getCategoryBySlug
      tags: [categories]
      summary: Get category by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{slug}/packs:
    get:
      operationId: getCategoryPacks
      tags: [categories]
      summary: Get packs in a category
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackListResponse"

  # Search
...
  /checkout/session:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
  /checkout/free:
    post:
      operationId: claimFreePack
      tags: [checkout]
      summary: Claim a free pack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimFreeBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidateResponse"

  # Admin endpoints
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        packCount:
          type: integer
        isFeatured:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name
        - packCount
        - isFeatured
        - sortOrder
        - createdAt
...
    Pack:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        shortDescription:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: ["string", "null"]
        categorySlug:
          type: ["string", "null"]
        status:
          type: string
        aiToolTargets:
          type: array
          items:
            type: string
        promptCount:
          type: integer
        priceCents:
          type: integer
        comparePriceCents:
          type: ["integer", "null"]
        isFree:
          type: boolean
        isFeatured:
          type: boolean
        isBestseller:
          type: boolean
        thumbnailUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        totalDownloads:
          type: integer
        avgRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        publishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - categoryId
        - status
        - aiToolTargets
        - promptCount
        - priceCents
        - isFree
        - isFeatured
        - isBestseller
        - tags
        - totalDownloads
        - reviewCount
        - createdAt
...
    PackListResponse:
      type: object
      properties:
        packs:
          type: array
          items:
            $ref: "#/components/schemas/Pack"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - packs
        - total
        - page
        - limit
        - totalPages
...
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, REFUNDED, DISPUTED]
        subtotalCents:
          type: integer
        discountCents:
          type: integer
        taxCents:
          type: integer
        totalCents:
          type: integer
        currency:
          type: string
        stripePaymentIntentId:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - id
        - userId
        - status
        - subtotalCents
        - discountCents
        - taxCents
        - totalCents
        - currency
        - createdAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        packId:
          type: integer
        packSlug:
          type: string
        titleSnapshot:
          type: string
        priceCents:
          type: integer
        downloadCount:
          type: integer
        firstDownloadedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - packId
        - packSlug
        - titleSnapshot
        - priceCents
        - downloadCount
...
    CreateCheckoutBody:
      type: object
      properties:
        packIds:
          type: array
          items:
            type: integer
        couponCode:
          type: string
      required:
        - packIds
    ClaimFreeBody:
      type: object
      properties:
        packId:
          type: integer
      required:
        - packId
    CheckoutSessionResponse:
      type: object
      properties:
        sessionUrl:
          type: string
        sessionId:
          type: string
        orderId:
          type: integer
        isMock:
          type: boolean
      required:
        - orderId
        - isMock

    # Coupon
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        subtotalCents:
          type: integer
      required:
        - code
    Coupon:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [PERCENT, FIXED]
        discountValue:
          type: number
        minOrderCents:
          type: ["integer", "null"]
        maxUses:
          type: ["integer", "null"]
        usesCount:
          type: integer
        validFrom:
          type: ["string", "null"]
          format: date-time
        validUntil:
          type: ["string", "null"]
          format: date-time
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - code
        - discountType
        - discountValue
        - usesCount
        - isActive
        - createdAt
    CouponValidateResponse:
      type: object
      properties:
        valid:
          type: boolean
        coupon:
          $ref: "#/components/schemas/Coupon"
        discountCents:
          type: integer
        message:
          type: string
      required:
        - valid
        - discountCents
        - message
...